+++ src/components/Form/example-mappers.js
@@ -1,24 +1,7 @@
 import React from 'react'
 import { TextInput } from '../index'
-
-function validateMinMaxFactory(
-  min = 0,
-  max = Number.MAX_SAFE_INTEGER,
-  message = `Must be between ${min} and ${max} characters`
-) {
-  function validateMinMax(value) {
-    const valueString = value ? String(value) : ''
-    const isInvalidMin = valueString && valueString.length < min
-    const isInvalidMax = valueString && valueString.length > max
-    return isInvalidMin || isInvalidMax ? message : undefined
-  }
-
-  return validateMinMax
-}
-
-function validateTruthy(x) {
-  return !!x ? '' : 'Please provide input'
-}
+import validateTruthy from '../../validators/validateTruthy'
+import validateMinMaxFactory from '../../validators/validateMinMax'
 
 function ValidatorGenerator(validatorName, args) {
   switch (validatorName) {
