+++ src/validators/validateTruthy.js
@@ -1,0 +1,3 @@
+export default function validateTruthy(x) {
+  return !!x ? '' : 'Please provide input'
+}
